refactor: add json extension when importing contract abis
