+++ templates/react/compound/packages/contracts/src/abis.js
@@ -1,22 +1,22 @@
-import COMP from "./abis/COMP";
-import base0bps_Slope2000bps from "./abis/base0bps_Slope2000bps";
-import base200bps_Slope222bps_Kink90_Jump10 from "./abis/base200bps_Slope222bps_Kink90_Jump10";
-import base200bps_Slope3000bps from "./abis/base200bps_Slope3000bps";
-import base500bps_Slope1200bps from "./abis/base500bps_Slope1200bps";
-import cBAT from "./abis/cBAT";
-import cDAI from "./abis/cDAI";
-import cETH from "./abis/cETH";
-import cREP from "./abis/cREP";
-import cSAI from "./abis/cSAI";
-import cTBTC from "./abis/cTBTC";
-import cUSDC from "./abis/cUSDC";
-import cWBTC from "./abis/cWBTC";
-import cZRX from "./abis/cZRX";
-import comptroller from "./abis/comptroller";
-import daiRateModel from "./abis/daiRateModel";
-import governance from "./abis/governance";
-import priceOracle from "./abis/priceOracle";
-import timelock from "./abis/timelock";
+import COMP from "./abis/COMP.json";
+import base0bps_Slope2000bps from "./abis/base0bps_Slope2000bps.json";
+import base200bps_Slope222bps_Kink90_Jump10 from "./abis/base200bps_Slope222bps_Kink90_Jump10.json";
+import base200bps_Slope3000bps from "./abis/base200bps_Slope3000bps.json";
+import base500bps_Slope1200bps from "./abis/base500bps_Slope1200bps.json";
+import cBAT from "./abis/cBAT.json";
+import cDAI from "./abis/cDAI.json";
+import cETH from "./abis/cETH.json";
+import cREP from "./abis/cREP.json";
+import cSAI from "./abis/cSAI.json";
+import cTBTC from "./abis/cTBTC.json";
+import cUSDC from "./abis/cUSDC.json";
+import cWBTC from "./abis/cWBTC.json";
+import cZRX from "./abis/cZRX.json";
+import comptroller from "./abis/comptroller.json";
+import daiRateModel from "./abis/daiRateModel.json";
+import governance from "./abis/governance.json";
+import priceOracle from "./abis/priceOracle.json";
+import timelock from "./abis/timelock.json";
 
 export default {
   base0bps_Slope2000bps,
