+++ templates/react/default/packages/contracts/src/abis.js
@@ -1,5 +1,5 @@
-import erc20Abi from "./abis/erc20";
-import ownableAbi from "./abis/ownable";
+import erc20Abi from "./abis/erc20.json";
+import ownableAbi from "./abis/ownable.json";
 
 const abis = {
   erc20: erc20Abi,
