+++ templates/react/default/packages/contracts/src/index.js
@@ -1,2 +1,2 @@
-export { default as abis } from "./abis.js";
-export { default as addresses } from "./addresses.js";
+export { default as abis } from "./abis";
+export { default as addresses } from "./addresses";
