+++ templates/react/uniswap-v1/packages/react-app/src/App.js
@@ -48,6 +48,7 @@ function App() {
         <p>
           Edit <code>packages/react-app/src/App.js</code> and save to reload.
         </p>
+        {/* Remove the "display: none" style and open the JavaScript console in the browser to see what this function does */}
         <button onClick={() => readOnChainData()} style={{ display: "none" }}>
           Read On-Chain Exchange Rate
         </button>
